+++ src/app/advertiser/edit-campaign/edit-campaign-summary/edit-campaign-summary.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { ActivatedRoute } from '@angular/router';
 import 'rxjs/add/operator/first';
 
 import { AppState } from 'models/app-state.model';
@@ -8,6 +7,7 @@ import { Campaign } from 'models/campaign.model';
 import { campaignStatusesEnum } from 'models/enum/campaign.enum';
 import { AdvertiserService } from 'advertiser/advertiser.service';
 import { AssetHelpersService } from 'common/asset-helpers.service';
+import { processTargeting } from 'common/components/targeting/targeting.helpers';
 import { adStatusesEnum } from 'models/enum/ad.enum';
 import { AddCampaignToCampaigns } from 'store/advertiser/advertiser.actions';
 import { HandleSubscription } from 'common/handle-subscription';
@@ -21,14 +21,13 @@ import { cloneDeep } from 'common/utilities/helpers';
 })
 export class EditCampaignSummaryComponent extends HandleSubscription implements OnInit {
   campaign: Campaign;
-  targetingOptionsToAdd: TargetingOption[];
-  targetingOptionsToExclude: TargetingOption[];
+  targetingOptionsToAdd: TargetingOption[] = [];
+  targetingOptionsToExclude: TargetingOption[] = [];
 
   constructor(
     private store: Store<AppState>,
     private advertiserService: AdvertiserService,
     private assetHelpers: AssetHelpersService,
-    private route: ActivatedRoute,
   ) {
     super();
   }
@@ -36,14 +35,18 @@ export class EditCampaignSummaryComponent extends HandleSubscription implements 
   ngOnInit(): void {
     const lastCampaignSubscription = this.store.select('state', 'advertiser', 'lastEditedCampaign')
       .first()
-      .subscribe((campaign: Campaign) => {
+      .subscribe(campaign => {
         this.assetHelpers.redirectIfNameNotFilled(campaign);
         this.campaign = campaign;
+        const targetingSubscription = this.advertiserService.getMedium(campaign.basicInformation.medium, campaign.basicInformation.vendor)
+          .take(1)
+          .subscribe(medium => {
+            this.targetingOptionsToAdd = processTargeting(medium)
+            this.targetingOptionsToExclude = cloneDeep(this.targetingOptionsToAdd)
+          })
+        this.subscriptions.push(targetingSubscription)
       });
     this.subscriptions.push(lastCampaignSubscription);
-
-    this.targetingOptionsToAdd = cloneDeep(this.route.parent.snapshot.data.targetingOptions);
-    this.targetingOptionsToExclude = cloneDeep(this.route.parent.snapshot.data.targetingOptions);
   }
 
   saveCampaign(isDraft): void {