+++ src/app/models/enum/ad.enum.ts
@@ -3,49 +3,11 @@ export enum adUnitTypesEnum {
   POP = 'pop',
 }
 
-export enum adTypesEnum {
-  IMAGE,
-  HTML,
-  DIRECT,
-  VIDEO,
-}
-
-export enum displayAdSizesEnum {
-  '300x250',
-  '336x280',
-  '728x90',
-  '300x600',
-  '320x100',
-  '320x50',
-  '468x60',
-  '234x60',
-  '120x600',
-  '120x240',
-  '160x600',
-  '300x1050',
-  '970x90',
-  '970x250',
-  '250x250',
-  '200x200',
-  '180x150',
-  '125x125',
-  '240x400',
-  '980x120',
-  '250x360',
-  '930x180',
-  '580x400',
-  '750x100',
-  '750x200',
-  '750x300',
-  '300x100',
-  '120x90',
-  '120x60',
-  '88x31',
-}
-
-export enum popAdSizesEnum {
-  'pop-up',
-  'pop-under',
+export const adCreativeTypes = {
+  IMAGE: 'image',
+  HTML: 'html',
+  DIRECT: 'direct',
+  VIDEO: 'video',
 }
 
 export enum adStatusesEnum {
@@ -55,27 +17,24 @@ export enum adStatusesEnum {
   REJECTED,
 }
 
-export enum validImageTypes {
-  'image/jpeg',
-  'image/jpg',
-  'image/pjpeg',
-  'image/png',
-  'image/gif',
-}
-
-export enum validHtmlTypes {
-  'application/zip',
-  'application/x-compressed',
-  'multipart/x-zip',
-  'application/octet-stream',
-  'application/x-zip-compressed',
-}
-
-export enum validVideoTypes {
-  'video/mp4',
-}
-
 export enum adUnitStatusesEnum {
   DRAFT,
   ACTIVE,
+}
+
+export const validHtmlTypes = [
+  'application/octet-stream',
+  'application/x-compressed',
+  'application/x-zip',
+  'application/x-zip-compressed',
+  'application/zip',
+  'multipart/x-zip',
+]
+
+export const fileTypes = {
+  'image/gif': 'GIF',
+  'image/jpeg': 'JPG',
+  'image/png': 'PNG',
+  'image/webp': 'WEBP',
+  'video/mp4': 'MP4',
 }