+++ src/app/publisher/classifier/classifier-list-item/classifier-list-item.component.ts
@@ -5,7 +5,6 @@ import { MatDialog } from '@angular/material'
 import { PublisherService } from 'publisher/publisher.service'
 import { BannerClassification } from 'models/classifier.model'
 import { HTTP_INTERNAL_SERVER_ERROR } from 'common/utilities/codes'
-import { adTypesEnum } from 'models/enum/ad.enum'
 import { ErrorResponseDialogComponent } from 'common/dialog/error-response-dialog/error-response-dialog.component'
 
 @Component({
@@ -19,7 +18,6 @@ export class ClassifierListItemComponent implements OnInit {
   readonly APPROVED: boolean = true
   readonly REJECTED: boolean = false
 
-  adTypesEnum = adTypesEnum
   isGlobal: boolean
 
   constructor (private publisherService: PublisherService, private dialog: MatDialog) {
@@ -27,10 +25,6 @@ export class ClassifierListItemComponent implements OnInit {
 
   ngOnInit (): void {
     this.isGlobal = this.siteId === null
-    this.bannerClassification = {
-      ...this.bannerClassification,
-      type: this.adTypesEnum[this.bannerClassification.type.toUpperCase()],
-    }
   }
 
   setClassificationStatus (isApproved: boolean | null): void {
@@ -52,7 +46,7 @@ export class ClassifierListItemComponent implements OnInit {
     this.setClassificationStatus(isApproved)
 
     this.publisherService.setBannerClassification(this.bannerClassification.bannerId, isApproved, this.siteId).
-      subscribe(() => { },
+      subscribe(() => { /* This is intentional */ },
         (error: HttpErrorResponse) => {
           this.setClassificationStatus(previousClassified)
           if (error.status !== HTTP_INTERNAL_SERVER_ERROR) {