+++ src/app/publisher/publisher-routing.module.ts
@@ -15,10 +15,10 @@ import {EditSiteSummaryComponent} from './edit-site/edit-site-summary/edit-site-
 import {PublisherGuard} from './publisher-guard.service';
 import {SiteResolver} from './resolvers/site.resolver';
 import {FilteringCriteriaResolver} from './resolvers/filtering-criteria.resolver';
+import {MediaResolver} from 'common/resolvers/media.resolver';
 import {AdUnitSizesResolver} from './resolvers/ad-unit-sizes.resolver';
 import {MatchingBannerSizesResolver} from "publisher/resolvers/matching-banner-sizes.resolver";
 import {LanguagesListResolver} from "publisher/resolvers/languages-list.resolver";
-import {TargetingCriteriaResolver} from 'publisher/resolvers/targeting-criteria.resolver'
 
 const publisherRoutes: Routes = [
   {
@@ -40,6 +40,9 @@ const publisherRoutes: Routes = [
             path: '',
             pathMatch: 'full',
             component: SiteDetailsComponent,
+            resolve: {
+              media: MediaResolver,
+            }
           },
           {
             path: 'classifier',
@@ -58,7 +61,7 @@ const publisherRoutes: Routes = [
             component: EditSiteBasicInformationComponent,
             resolve: {
               languagesList: LanguagesListResolver,
-              targetingOptions: TargetingCriteriaResolver,
+              media: MediaResolver,
             }
           },
           {
@@ -92,7 +95,10 @@ const publisherRoutes: Routes = [
           {
             path: 'basic-information',
             component: EditSiteBasicInformationComponent,
-            resolve: {languagesList: LanguagesListResolver}
+            resolve: {
+              languagesList: LanguagesListResolver,
+              media: MediaResolver,
+            }
           },
           {
             path: 'pops-settings',