+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Targeting supports taxonomy v4
 ### Fixed
 - Video preview overflow
+- Pages order while creating site (navigation)
+- Targeting option indicator while removing complex option
 
 ## [1.14.1] - 2022-02-16
 ### Fixed